--- lerna.json
@@ -1,6 +1,0 @@
-{
-  "packages": ["packages/*"],
-  "npmClient": "npm",
-  "useWorkspaces": true,
-  "version": "independent"
-}
