+++ package.json
@@ -1,27 +1,42 @@
 {
-  "name": "@digicat/ui-component-library",
+  "name": "@digicatapult/ui-component-library",
   "version": "0.1.3",
   "private": true,
   "description": "UI Component Library",
   "author": "Digital Catapult (https://www.digicatapult.org.uk)",
   "license": "Apache-2.0",
   "type": "module",
-  "workspaces": [
-    "packages/*",
-    "storybook"
-  ],
   "scripts": {
-    "bootstrap": "lerna bootstrap --use-workspaces",
-    "build": "lerna exec --ignore storybook --parallel -- tsc",
-    "build:watch": "lerna exec --ignore storybook --parallel -- tsc --watch",
+    "build": "tsc",
+    "build:watch": "tsc --watch",
     "coverage": "jest --coverage",
     "depcheck": "depcheck",
     "lint": "eslint .",
+    "storybook": "start-storybook -p 9000",
     "test": "jest"
   },
+  "peerDependencies": {
+    "react": "^18.2.0"
+  },
+  "dependencies": {
+    "prop-types": "^15.8.1",
+    "styled-components": "^5.3.6"
+  },
   "devDependencies": {
+    "@storybook/addon-actions": "^6.5.14",
+    "@storybook/addon-essentials": "^6.5.14",
+    "@storybook/addon-links": "^6.5.14",
+    "@storybook/builder-webpack5": "^6.5.14",
+    "@storybook/manager-webpack5": "^6.5.14",
+    "@storybook/react": "^6.5.14",
+    "@storybook/source-loader": "^6.5.14",
     "@types/jest": "^29.2.4",
+    "@types/prop-types": "^15.7.5",
     "@types/react": "^18.0.26",
+    "@types/react-test-renderer": "^18.0.0",
+    "@types/styled-components": "^5.1.26",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "@typescript-eslint/parser": "^5.46.0",
     "core-js": "^3.26.1",
     "depcheck": "^1.4.3",
     "eslint": "^8.29.0",
@@ -34,8 +49,12 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-workspaces": "^0.7.0",
     "jest": "^29.3.1",
-    "lerna": "^6.1.0",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-styled-components": "^7.1.1",
     "prettier": "^2.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0",
     "ts-jest": "^29.0.3",
     "typescript": "^4.9.4"
   }
