--- packages/components/README.md
@@ -1,33 +1,0 @@
-<h1 align="center">DC Component Library</h1>
-
-React components built for Digital Catapult projects 🧠
-
-## Installation
-
-`@digicatapult/components` are available as an [npm package](https://www.npmjs.com/package/@digicatapult/components).
-
-```sh
-// with npm
-npm install @digicatapult/components
-
-// with yarn
-yarn add @digicatapult/components
-```
-
-## Usage
-
-```jsx
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Button } from '@digicatapult/components'
-
-function App() {
-  return <Button>hello</Button>
-}
-
-ReactDOM.render(<App />, document.querySelector('#app'))
-```
-
-## Documentation
-
-Check out our [storybook](https://digicat-components.netlify.app/) for documentation of individual components
