--- packages/components/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@digicatapult/components",
-  "version": "0.1.2",
-  "description": "React components for Digital Catapult projects.",
-  "license": "Apache-2.0",
-  "modules": "lib/index.js",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/digicatapult/ui-component-library.git"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "prop-types": "^15.8.1",
-    "styled-components": "^5.3.6"
-  },
-  "peerDependencies": {
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react-test-renderer": "^18.0.0",
-    "jest-styled-components": "^7.1.1",
-    "react": "^18.2.0",
-    "react-test-renderer": "^18.2.0"
-  }
-}
