--- packages/components/src/TextArea/index.js
@@ -1,1 +1,0 @@
-export { default } from './TextArea'
