--- packages/components/src/index.js
@@ -1,5 +1,0 @@
-export { default as Button } from './Button'
-export { default as Input } from './Input'
-export { default as TextArea } from './TextArea'
-export { default as Card } from './Card'
-export { default as TestComponent } from './TestComponent'
