+++ src/Button/index.js
@@ -1,0 +1,2 @@
+import Button from './Button.js'
+export default Button
