+++ src/Card/Card.spec.js
@@ -2,7 +2,7 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 
-import Card from '.'
+import Card from './index.js'
 
 const date = new Date('December 17, 1995 03:24:00')
 
@@ -14,6 +14,7 @@ describe('Card', () => {
           title="Card story"
           name="Story Writer"
           date={date}
+          locale="en-US"
           description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Est ullamcorper eget nulla facilisi etiam dignissim diam quis. Sem et tortor consequat id porta. Viverra nam libero justo laoreet sit."
         />
       )
