+++ src/Card/index.js
@@ -1,0 +1,2 @@
+import Card from './Card.js'
+export default Card
