+++ src/Input/index.js
@@ -1,0 +1,2 @@
+import Input from './Input.js'
+export default Input
