+++ src/TextArea/index.js
@@ -1,0 +1,2 @@
+import TextArea from './TextArea.js'
+export default TextArea
