+++ src/index.js
@@ -1,0 +1,5 @@
+export { default as Button } from './Button/index.js'
+export { default as Input } from './Input/index.js'
+export { default as TextArea } from './TextArea/index.js'
+export { default as Card } from './Card/index.js'
+export { default as TestComponent } from './TestComponent/index.tsx'
